requirements.txt
streamlit
tensorflow
keras
numpy
Pillow
huggingface_hub 